Refactor from assB
