+++ src/main/java/VotingApp/IoTDevice.java
