+++ src/main/java/VotingApp/Notification.java
