+++ src/main/java/VotingApp/Poll.java
