+++ src/main/java/VotingApp/Vote.java
